I chose the "Folder based" way and rename "landing" in "index"
